terminal v1.0.1

+++ src/app/shared/system-service/state.management.service.ts
@@ -24,12 +24,12 @@ export class StateManagmentService{
     }
 
     hasState(id:number):boolean{
-
         return this._sessionStateManagmentService.has(id) ? true : false;
     }
 
     removeState(id:number ): void{
-        this._sessionStateManagmentService.delete(id)
+       if(this._sessionStateManagmentService.has(id))
+            this._sessionStateManagmentService.delete(id)
     }
 
     getKeys():number[]{
